requirements.txt
python-dotenv
onnxruntime
numpy
pillow
